+++ src/types/observableobject.ts
@@ -5,9 +5,17 @@
  */
 import ObservableValue from "./observablevalue";
 import ComputedValue from "../core/computedvalue";
-import {ValueMode, makeChildObservable, AsStructure} from '../core';
-import {IObjectChange, Lambda} from '../interfaces';
+import {ValueMode, makeChildObservable, AsStructure} from './modifiers';
+import {Lambda} from '../utils/utils';
 import SimpleEventEmitter from '../utils/simpleeventemitter';
+
+
+export interface IObjectChange<T, R> {
+    name: string;
+    object: R;
+    type: string;
+    oldValue?: T;
+}
 
 // responsible for the administration of objects that have become reactive
 export class ObservableObject { // TODO: implement IObservable
@@ -82,4 +90,8 @@ export class ObservableObject { // TODO: implement IObservable
 	observe(callback: (changes:IObjectChange<any, any>) => void): Lambda {
 		return this._events.on(callback);
 	}
+}
+
+export function isObservableObject(thing):boolean {
+    return thing && typeof thing === "object" && thing.$mobservable instanceof ObservableObject;
 }