+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This means that it is allowed in an to change the state in such a way that the s
 * Changed error handling: Mobservable will always fail fast now. This means that a reactive computation should never allow an exception to escape as Mobservable will no longer try to recover from it
 * `.fastArray` has been deprecated and has become the default when creating arrays.
 * Mobservable is now extensible! Use the exposed `Atom` or `Reaction` classes to define your own custom state values and reactive actions.
+* Removed deprecated methods: `isReactive`, `makeReactive`, `observeUntil`, `observeAsync`
 
 # 1.2.4